move queries and mutations to dedicated layer
