+++ src/app/games/[id]/scoreEntry.tsx
@@ -6,7 +6,7 @@ import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 
-import createScoresForGame from "./createScoresForGameAction";
+import { createScoresForGame } from "@/server/mutations";
 
 import { GameWithPlayersAndScores } from "./page";
 import { Player } from "./page";
